pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wifi-mapping"
version = "0.1.0"
description = "WiFi Signal Triangulation & 3D Indoor Mapping"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "pyarrow>=12.0",
    "pyserial>=3.5",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "plotly>=5.15",
    "streamlit>=1.30",
]

[project.optional-dependencies]
advanced = [
    "open3d>=0.17",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "csikit>=0.2",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]